final-pjt-back/requirements.txt
asgiref==3.4.1
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.9
django-cors-headers==3.5.0
django-extensions==3.1.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
ipython==7.29.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.21.4
pandas==1.3.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.22
Pygments==2.10.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
wcwidth==0.2.5
